app: Don't modify XDG_DATA_DIRS environment variable
On start-up the Mission Center modifies the XDG_DATA_DIRS variable by adding
`$HOME/.local/share` to it, as well as adding some common paths in case it is
empty or unset.

This was done as a workaround for some systems that were missconfigured. Since
all DEs should set this variable up correctly remove this workaround.