services: Use a hashed dbus object path as a unique service id
Magpie now supplies a `u64` as a service identifier instead of using the
service name. The name is now a separate field of the `Service` type.

The code adapts to the changes by making all functions and data structures
now use u64 as the service identifier.

Moreover a new field is added to the `RowModel` GLib type for the service id.