services: Try to fix enabling and disabling services
Retain the service state when listing all available units during a full
refresh. This state is then assigned, to each service, when updating
individual service properties.

This has the side effect that service state is only updated when a full refresh
occurs. To try to make sure that the service state is correct, force a full
refresh when enabling or disabling a service.