services: Fix right-clicking any item would select the first item in the list
When right-clicking, to open the context menu, any item in the list said item
should be selected automatically.

The logic to select an item is dependent on the item having a valid `id`. When
switching from the service name to a unique `u64` for the service id this
property was no longer set, which caused the selection logic to fail.

The fix is to set the `id` property to the string-representation of the service
id.

See merge request mission-center-devs/mission-center!419